classic: use ABS_X and ABS_Y so udev tags as joystick
My main motivation is SDL2 which has udev and hotplug support
(sweet!) and everything else you could ever want, and expects your
joysticks to be labeled as joysticks. See SDL_sysjoystick.c:

    UDEV_udev_enumerate_add_match_property(enumerate, "ID_INPUT_JOYSTICK", "1");

This is great except for our classic controller, which is consider not
a joystick. From udev-builtin-input_id.c (now in the systemd tree):

        if (test_bit (EV_ABS, bitmask_ev) &&
            test_bit (ABS_X, bitmask_abs) && test_bit (ABS_Y, bitmask_abs)) {
                [ ... ]
                else if (test_bit (BTN_TRIGGER, bitmask_key) ||
                         test_bit (BTN_A, bitmask_key) ||
                         test_bit (BTN_1, bitmask_key))
                        udev_builtin_add_property(dev, test, "ID_INPUT_JOYSTICK", "1");